added new class WindowExamples.cs (extracted tests from ApiExamples.cs)
