+++ src/Coypu.AcceptanceTests/Examples/ApiExamples.cs
@@ -15,16 +15,6 @@ namespace Coypu.AcceptanceTests.Examples
     [TestFixture]
     public class ApiExamples : WaitAndRetryExamples
     {
-        private object GetOuterHeight()
-        {
-            return Browser.ExecuteScript("return window.outerHeight;");
-        }
-
-        private object GetOuterWidth()
-        {
-            return Browser.ExecuteScript("return window.outerWidth;");
-        }
-
         public class CustomFirefoxProfileSeleniumWebDriver : SeleniumWebDriver
         {
             public CustomFirefoxProfileSeleniumWebDriver(Browser browser) : base(CustomOptions(), browser) { }
@@ -146,18 +136,7 @@ namespace Coypu.AcceptanceTests.Examples
                                .Text,
                         Is.EqualTo("Hover on me - hovered"));
         }
-
-        [Test]
-        public void MaximiseWindow()
-        {
-            var availWidth = Browser.ExecuteScript("return window.screen.availWidth;");
-            var initalWidth = GetOuterWidth();
-            Assert.That(initalWidth, Is.LessThan(availWidth));
-
-            Browser.MaximiseWindow();
-            Assert.That(GetOuterWidth(), Is.GreaterThanOrEqualTo(availWidth));
-        }
-
+        
         [Test]
         public void ModalDialog_while_multiple_windows_are_open()
         {
@@ -188,28 +167,6 @@ namespace Coypu.AcceptanceTests.Examples
             Assert.That(Browser.FindButton("clickMeTest")
                                .Value,
                         Is.EqualTo("Click me - clicked"));
-        }
-
-        [Test]
-        public void RefreshingWindow()
-        {
-            var tickBeforeRefresh = (long) Browser.ExecuteScript("return window.SpecData.CurrentTick;");
-            Browser.Refresh();
-            var tickAfterRefresh = (long) Browser.ExecuteScript("return window.SpecData.CurrentTick;");
-            Assert.That(tickAfterRefresh - tickBeforeRefresh, Is.GreaterThan(0));
-        }
-
-        [Test]
-        public void ResizeWindow()
-        {
-            var initalWidth = GetOuterWidth();
-            var initialHeight = GetOuterHeight();
-            Assert.That(initalWidth, Is.Not.EqualTo(500));
-            Assert.That(initialHeight, Is.Not.EqualTo(600));
-
-            Browser.ResizeTo(500, 600);
-            Assert.That(GetOuterWidth(), Is.EqualTo(500));
-            Assert.That(GetOuterHeight(), Is.EqualTo(600));
         }
 
         [Test]
@@ -298,24 +255,6 @@ namespace Coypu.AcceptanceTests.Examples
                              () => Browser.HasContent("try until 5"),
                              TimeSpan.FromMilliseconds(50),
                              new Options {Timeout = TimeSpan.FromMilliseconds(10000)});
-        }
-
-        [Test]
-        public void WindowScoping_example()
-        {
-            var mainWindow = Browser;
-            Assert.That(mainWindow.FindButton("scoped button", Options.First)
-                                  .Id,
-                        Is.EqualTo("scope1ButtonId"));
-            mainWindow.ExecuteScript("setTimeout(function() {document.getElementById(\"openPopupLink\").click();}), 3000");
-            var popUp = mainWindow.FindWindow("Pop Up Window");
-            Assert.That(popUp.FindButton("scoped button")
-                             .Id,
-                        Is.EqualTo("popUpButtonId"));
-            // And back
-            Assert.That(mainWindow.FindButton("scoped button", Options.First)
-                                  .Id,
-                        Is.EqualTo("scope1ButtonId"));
         }
     }
 }
